Add tract point in poly operation
