+++ src/main/resources/application.conf
@@ -16,7 +16,7 @@ hmda {
 database = {
   dataSourceClass = "org.postgresql.ds.PGSimpleDataSource"
   properties = {
-    databaseName = ${?PG_DATABASE_NAME}
+    databaseName = ${?PG_DATABASE}
     user = ${?PG_USER}
     password = ${?PG_PASSWORD}
   }
